pyproject.toml
[project]
name = "spookycat"
version = "0.1.0"
description = "Custom Elgato Stream Deck Mini controller"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.2.0",
    "streamdeck>=0.9.8",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
]
